"tree" -> Pipeline converter, very rough. Also added toTree to MappedClosure
Chances are we'll end up calling toTree a couple times...or we'll end
up junking it. We shall see. I think it'll mainly be relevant in
translating Phases into code that we can then wrap in parallel blocks etc.

The converter is actually straight-up wrong right now. It's just
"tree" to Groovy closure, not Pipeline script - there's a lot more
logic that's needed for transforming the model classes to Pipeline,
but this helped me get my head around where I was going, so hey.