A whole bunch of work.
* PlumberDependencyGraph and model class stuff generally works.
* Some tests for those, especially the graph
* A lot of attempts to get the step to work with closures as args.

Closures as args turned out to have a huge problem - they're
transformed to CpsClosure2s, and CpsClosure2 extends CpsClosure...and
CpsClosure overrides Closure.call(...) to do magic with throwable
invocation stuff. The result is that you can't use normal Groovy
Closure behavior on any closure that's declared inside a Pipeline
script (Jenkinsfile, global library, contributed global
variable/script, whatever) unless that closure is declared inside a
@NonCPS method.

So for the moment, I'm bypassing that issue by cheating hard and
passing the closure in as a string and then evaluating it into a
closure so I can work on other things farther along the flow.