Actually functional code-gen.
It's not pretty yet, it's kind of demented, but it *does* generally
work. I'm probably missing places that need quotes and the like, and
I'd rather be writing the code-gen stuff a bit better than "add this
line to this list of lines and then JOIN THEM ALL UP" but it *does*
work. So that's interesting, yes?